Allow setting a real voltage, rather than a DAC input code The current `setVoltage` method is badly named, because the caller is not actually setting a voltage, they’re setting a DAC input code.  Since the library has already been released, we can’t now change the name of this method or we’ll lose backwards compatibility, so I have had to name the new method `setNearestActualVoltage`.  It would have been clearer to have e.g. `setDacInputCode` and `setNearestVoltage`.
I suspect that a lot (most?) users of this library will want to be able to set an actual voltage, rather than have to calculate the input code required to generate the voltage they want.  For those that do want to set a voltage, this new method makes client code a lot cleaner.  Clearly, in order to do this, we need to know the reference voltage, so this is passed in to the new method.

`setNearestActualVoltage` returns a `float` which represents the nominal output voltage that is nearest to the requested value.